+++ src/components/mode-toggler.tsx
@@ -1,14 +1,16 @@
 import { Moon, Sun } from "lucide-react"
 import { Button } from "@/components/ui/button"
 import { useTheme } from "./theme-provider"
+import { useTranslation } from "react-i18next"
 
 export function ModeToggle() {
   const {resolvedTheme, setTheme} = useTheme()
+  const {t} = useTranslation()
   return (
     <Button
       className="shadow-xs text-primary rounded-r-4xl"
       variant="ghost"
-      size="icon" title="Փոխել ոճը"
+      size="icon" title={t("theme")}
       onClick={()=>setTheme(resolvedTheme==="dark" ? "light" : "dark")}
     >
       {resolvedTheme==="dark" ? (