+++ src/i18n/config.ts
@@ -1,0 +1,8 @@
+import { languageOption } from "./types";
+
+export const languages: languageOption[] = [
+     { language: "English", code: "en", countryCode: "us" },
+     { language: "Հայերեն", code: "hy", countryCode: "am" },
+     { language: "Русский", code: "ru", countryCode: "ru" },
+     { language: "ქართული", code: "ka", countryCode: "ge" }
+];