+++ src/i18n/types.ts
@@ -1,0 +1,3 @@
+export type LangCode = "en" | "hy" | "ru" | "ka"
+export type CountryCode = "us" | "am" | "ru" | "ge"
+export type languageOption = { language: string; code: LangCode, countryCode: CountryCode };