+++ src/icons/night.tsx
@@ -1,0 +1,11 @@
+import SVGWrapper from ".";
+
+export default function Night(props: React.SVGProps<SVGElement>) {
+     return (
+          <SVGWrapper {...props} viewBox="0 0 24 24">
+               <path
+                    d="M12.1 22q-2.1 0-3.937-.8t-3.2-2.162t-2.163-3.2T2 11.9q0-3.65 2.325-6.437T10.25 2q-.45 2.475.275 4.838t2.5 4.137t4.138 2.5T22 13.75q-.65 3.6-3.45 5.925T12.1 22"
+               />
+          </SVGWrapper>
+     )
+}