+++ src/lib/types/index.ts
@@ -1,0 +1,13 @@
+import { SoundsType } from "../sounds";
+
+export interface ISounds {
+      id: SoundsType;
+      url: string;
+      Icon: React.ComponentType<React.SVGProps<SVGSVGElement>>;
+}
+export type PlayBackType = "idle" | "playing" | "paused"
+export enum NavLinks{
+      Sounds = 'sounds',
+      BreathingExercise = 'exercise',
+      Pomodoro = 'pomodoro'
+}