+++ src/main.tsx
@@ -4,6 +4,7 @@ import { ThemeProvider } from "./components/theme-provider";
 import { RouterProvider } from "react-router";
 import { router } from "./routes";
 import "./App.css";
+import "@/i18n"
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
   <React.StrictMode>