+++ src/pages/sounds.tsx
@@ -1,8 +1,10 @@
 import WindowWrapper from "@/components/window";
+import { useTranslation } from "react-i18next";
 
 export default function SoundsPage(){
+     const {t} = useTranslation("sounds")
      return (
-          <WindowWrapper title="Հանգստացնող ձայններ">
+          <WindowWrapper title={t("title")}>
                Hello New Calm World, Goodbye Hellish world!
           </WindowWrapper>
      )